+++ tests/test_save_load_team.cpp
@@ -368,7 +368,7 @@ static int load_menu_injector(void* data)
         interact_match("back", [](const Interactable& item) { return item.y >= 170; });
     }
 
-    const Uint32 deadline = SDL_GetTicks() + 10000;
+    const Uint64 deadline = SDL_GetTicks() + 10000;
     while (SDL_GetTicks() < deadline)
     {
         if (wait_for_interactable("continue_game", 150))
